import validate function, fixes logout style
